Fixed deleting previous data on type change
Named validator return tuple items for better readability
