+++ MageNet.Services/AttributeRepository/TypeBearers/PriceTypeBearer.cs
@@ -1,3 +1,4 @@
+using System.Globalization;
 using MageNet.Persistence.Exceptions;
 using MageNet.Persistence.Models.AbstractModels.ModelEnums;
 using MageNet.Persistence.Models.AbstractModels.ModelInterfaces;
@@ -5,17 +6,17 @@ using MageNet.Persistence.Models.Attributes;
 using MageNetServices.AttributeRepository.DTO.Attributes;
 using MageNetServices.Interfaces;
 
-namespace MageNetServices.AttributeRepository.DTO.TypeBearers;
+namespace MageNetServices.AttributeRepository.TypeBearers;
 
-public class SelectableTypeBearer : IAttributeTypeBearer
+public class PriceTypeBearer : IAttributeTypeBearer
 {
-    private readonly IAttributeDataRepository<SelectableAttributeData> _dataRepository;
+    private readonly IAttributeDataRepository<PriceAttributeData> _dataRepository;
 
 
-    public SelectableTypeBearer(IAttributeDataRepository<SelectableAttributeData> dataRepository)
+    public PriceTypeBearer(IAttributeDataRepository<PriceAttributeData> dataRepository)
     {
         _dataRepository = dataRepository;
-        AttributeType = AttributeType.Selectable;
+        AttributeType = AttributeType.Price;
     }
 
     public AttributeType AttributeType { get; }
@@ -27,7 +28,6 @@ public class SelectableTypeBearer : IAttributeTypeBearer
 
         var attributeId = _dataRepository.CreateAttribute(attributeEntity);
         attributeData.AttributeId = attributeId;
-
         _dataRepository.CreateAttributeData(attributeData);
         _dataRepository.SaveChanges();
 
@@ -44,9 +44,9 @@ public class SelectableTypeBearer : IAttributeTypeBearer
             EntityId = attribute.EntityId,
             AttributeName = attribute.AttributeName,
             AttributeType = attribute.AttributeType.AttributeType,
-            DefaultLiteralValue = null,
-            SelectableOptions = attributeData.Values,
-            IsMultipleSelect = attributeData.IsMultipleSelect
+            DefaultLiteralValue = attributeData.DefaultValue.ToString(CultureInfo.CurrentCulture),
+            SelectableOptions = null,
+            IsMultipleSelect = null
         };
     }
 
@@ -70,7 +70,7 @@ public class SelectableTypeBearer : IAttributeTypeBearer
 
     public (IAttributeEntity, IAttributeData) DecoupleAttributeWithData(IAttributeWithData attributeWithData)
     {
-        var attribute = new AttributeEntity()
+        var attribute = new AttributeEntity
         {
             AttributeId = attributeWithData.AttributeId,
             AttributeName = attributeWithData.AttributeName,
@@ -78,22 +78,27 @@ public class SelectableTypeBearer : IAttributeTypeBearer
             EntityId = attributeWithData.EntityId,
         };
 
-        var attributeData = new SelectableAttributeData()
+        var attributeData = new PriceAttributeData
         {
             AttributeId = attributeWithData.AttributeId,
-            Values = attributeWithData.SelectableOptions ?? throw new SelectableOptionsNotProvidedException(
-                "Null selectable options are not allowed. Please send the empty list of options if you plan to add them later."),
-            IsMultipleSelect = attributeWithData.IsMultipleSelect ??
-                               throw new IsMultipleSelectValueNotProvidedException(
-                                   "IsMultipleSelect value must be provided for the selectable attributes.")
+            DefaultValue = decimal.Parse(s: attributeWithData.DefaultLiteralValue
+                                            ?? throw new DefaultLiteralValueNotProvidedException(
+                                                "Price attribute must have a decimal value"))
         };
 
         return (attribute, attributeData);
     }
 
+    public void RemoveDbData(Guid attributeId)
+    {
+        _dataRepository.DeleteAttributeData(attributeId);
+    }
+
     public IAttributeWithData RemoveIrrelevantProperties(IAttributeWithData attributeWithData)
     {
-        attributeWithData.DefaultLiteralValue = null;
+        attributeWithData.SelectableOptions = null;
+        attributeWithData.IsMultipleSelect = null;
+
         return attributeWithData;
     }
 }
