+++ MageNet.Services/AttributeRepository/TypeBearers/SelectableTypeBearer.cs
@@ -5,17 +5,17 @@ using MageNet.Persistence.Models.Attributes;
 using MageNetServices.AttributeRepository.DTO.Attributes;
 using MageNetServices.Interfaces;
 
-namespace MageNetServices.AttributeRepository.DTO.TypeBearers;
+namespace MageNetServices.AttributeRepository.TypeBearers;
 
-public class PriceTypeBearer : IAttributeTypeBearer
+public class SelectableTypeBearer : IAttributeTypeBearer
 {
-    private readonly IAttributeDataRepository<PriceAttributeData> _dataRepository;
+    private readonly IAttributeDataRepository<SelectableAttributeData> _dataRepository;
 
 
-    public PriceTypeBearer(IAttributeDataRepository<PriceAttributeData> dataRepository)
+    public SelectableTypeBearer(IAttributeDataRepository<SelectableAttributeData> dataRepository)
     {
         _dataRepository = dataRepository;
-        AttributeType = AttributeType.Price;
+        AttributeType = AttributeType.Selectable;
     }
 
     public AttributeType AttributeType { get; }
@@ -27,6 +27,7 @@ public class PriceTypeBearer : IAttributeTypeBearer
 
         var attributeId = _dataRepository.CreateAttribute(attributeEntity);
         attributeData.AttributeId = attributeId;
+
         _dataRepository.CreateAttributeData(attributeData);
         _dataRepository.SaveChanges();
 
@@ -43,9 +44,9 @@ public class PriceTypeBearer : IAttributeTypeBearer
             EntityId = attribute.EntityId,
             AttributeName = attribute.AttributeName,
             AttributeType = attribute.AttributeType.AttributeType,
-            DefaultLiteralValue = attributeData.DefaultValue.ToString(),
-            SelectableOptions = null,
-            IsMultipleSelect = null
+            DefaultLiteralValue = null,
+            SelectableOptions = attributeData.Values,
+            IsMultipleSelect = attributeData.IsMultipleSelect
         };
     }
 
@@ -69,7 +70,7 @@ public class PriceTypeBearer : IAttributeTypeBearer
 
     public (IAttributeEntity, IAttributeData) DecoupleAttributeWithData(IAttributeWithData attributeWithData)
     {
-        var attribute = new AttributeEntity
+        var attribute = new AttributeEntity()
         {
             AttributeId = attributeWithData.AttributeId,
             AttributeName = attributeWithData.AttributeName,
@@ -77,22 +78,27 @@ public class PriceTypeBearer : IAttributeTypeBearer
             EntityId = attributeWithData.EntityId,
         };
 
-        var attributeData = new PriceAttributeData
+        var attributeData = new SelectableAttributeData()
         {
             AttributeId = attributeWithData.AttributeId,
-            DefaultValue = decimal.Parse(s: attributeWithData.DefaultLiteralValue
-                                            ?? throw new DefaultLiteralValueNotProvidedException(
-                                                "Price attribute must have a decimal value"))
+            Values = attributeWithData.SelectableOptions ?? throw new SelectableOptionsNotProvidedException(
+                "Null selectable options are not allowed. Please send the empty list of options if you plan to add them later."),
+            IsMultipleSelect = attributeWithData.IsMultipleSelect ??
+                               throw new IsMultipleSelectValueNotProvidedException(
+                                   "IsMultipleSelect value must be provided for the selectable attributes.")
         };
 
         return (attribute, attributeData);
     }
 
+    public void RemoveDbData(Guid attributeId)
+    {
+        _dataRepository.DeleteAttributeData(attributeId);
+    }
+
     public IAttributeWithData RemoveIrrelevantProperties(IAttributeWithData attributeWithData)
     {
-        attributeWithData.SelectableOptions = null;
-        attributeWithData.IsMultipleSelect = null;
-
+        attributeWithData.DefaultLiteralValue = null;
         return attributeWithData;
     }
 }
