+++ MageNet.Services/AttributeRepository/TypedDataRepositories/TextAttributeDataRepo.cs
@@ -3,7 +3,7 @@ using MageNet.Persistence.Exceptions;
 using MageNet.Persistence.Models.AbstractModels.ModelInterfaces;
 using MageNet.Persistence.Models.Attributes;
 
-namespace MageNetServices.AttributeRepository.DTO.TypedDataRepositories;
+namespace MageNetServices.AttributeRepository.TypedDataRepositories;
 
 public class TextAttributeDataRepo : AttributeDataRepo<TextAttributeData>
 {
@@ -20,19 +20,24 @@ public class TextAttributeDataRepo : AttributeDataRepo<TextAttributeData>
 
         return textAttributeData ?? throw new AttributeNotFoundException();
     }
-    
+
 
     public override void CreateAttributeData(IAttributeData attributeData)
     {
         _dbContext.TextAttributes.Add(attributeData as TextAttributeData
-                                       ?? throw new InvalidAttributeDataException(
-                                           $"Attribute Data cannot be safely casted to Text Attribute Data"));
+                                      ?? throw new InvalidAttributeDataException(
+                                          $"Attribute Data cannot be safely casted to Text Attribute Data"));
     }
 
     public override void UpdateAttributeData(IAttributeData attributeData)
     {
         _dbContext.TextAttributes.Update(attributeData as TextAttributeData ??
-                                          throw new InvalidOperationException(
-                                              $"Attribute Data cannot be safely casted to Text Attribute Data"));
+                                         throw new InvalidOperationException(
+                                             $"Attribute Data cannot be safely casted to Text Attribute Data"));
+    }
+
+    public override void DeleteAttributeData(Guid attributeId)
+    {
+        _dbContext.TextAttributes.Remove(_dbContext.TextAttributes.Single(x => x.AttributeId == attributeId));
     }
 }
