+++ MageNet.Services/Interfaces/IAttributeValidator.cs
@@ -4,5 +4,6 @@ namespace MageNetServices.Interfaces;
 
 public interface IAttributeValidator
 {
-    public (bool, IEnumerable<ValidationException>) CheckAttributeValidity(IAttributeWithData attributeWithData);
+    public (bool isValid, IEnumerable<ValidationException> validationErrors) CheckAttributeValidity(
+        IAttributeWithData attributeWithData);
 }
